requirements.txt
flask>=2.2.0
openai>=1.3.7
python-dotenv>=1.0.0
requests>=2.31.0
gunicorn>=20.1.0
waitress>=2.1.0
markdown
